analizer-DDoS/requirements.txt
scapy
requests
python-dotenv
streamlit
plotly
pandas
scikit-learn
joblib
numpy